backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
celery>=5.4.0
redis>=5.0.0
pyomo>=6.7.0
pandas>=2.2.0
highspy>=1.8.0
openpyxl>=3.1.0
matplotlib>=3.9.0
pytest>=8.0.0
